pyproject.toml
[tool.poetry]
name = "lte-monitoring-campaign"
version = "0.1.0"
description = "This repository contains code for establishing monitoring campaign for LTE technologies like NB-IoT and LTE-M. We compute PDR,throughput and energy consume"
authors = ["cafekrem <dutriezcle@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
SQLAlchemy = "^2.0.29"
influxdb = "^5.3.1"
influxdb-client = "^1.42.0"
pandas = "^2.2.2"
mysql-connector-python = "^8.3.0"
mysqlclient = "^2.2.4"
jinja2 = "^3.1.3"
hydra-core = "^1.3.2"
pyvisa = "^1.14.1"
paho-mqtt = "^2.0.0"
visa = "^1.0.0"
pyvisa-py = "^0.7.2"
pyusb = "^1.2.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
